requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
google-auth==2.40.3
stripe==12.0.0
